+++ Assets/Scripts/Hugger.cs
@@ -9,7 +9,7 @@ public class Hugger : Enemy
 
     public int timesPlayerDidIt;
     bool grabbing;
-    int damage;
+    float damage;
 
     // Start is called before the first frame update
     void Start()
